networkApi/requirements.txt
django
djangorestframework
django-cors-headers
networkx